Implement Chase Lev Deque for work stealing queue
Implement the work stealing queue described in Dynamic circular work-stealing deque (Chase, Lev 2005), with some quirks and extensions required by MPL.
The specific memory annotations are taken from Correct and efficient work-stealing for weak memory models (https://doi.org/10.1145/2442516.2442524, PPoPP 2013) with minor modifications.

First quirk is that this implementation actually is NOT a dynamic array.
The vast majority of workloads never rack up more than 64 tasks, so the array is fixed at 64.

It also includes a `setDepth` operation, which is a hack that should hopefully be addressed later.